libcamera: pipeline: vivid: Buffer handling and stream control
We can now add buffer management, and connect up our bufferReady signal
to a callback.

Note that we provide the ability to export buffers from our capture
device (data->video_) using the exportBuffers() functionality from the
V4L2VideoDevice which allows a FrameBufferAllocater to obtain buffers
from this device.

When buffers are obtained through the exportFrameBuffers API, they are
orphaned and left unassociated with the device, and must be reimported
at start() time anyway. This allows the same interface to be used
whether internal buffers, or external buffers are used for the stream.

When a buffer completes, we call the buffer completion handler on the
pipeline handler, and because we have only a single stream, we can also
immediately complete the request.

Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>