libcamera: pipeline: vivid: Queue requests
When a request is given to a pipeline handler, it must parse the request
and identify what actions the pipeline handler should take to enact on hardware.

In the case of the VIVID pipeline handler, we identify the buffer from the only
supported stream, and queue it to the video capture device.

Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>