DocumentationUseOnly: Force enable Vivid
To make sure we always build the Vivid pipeline in this branch,
explicitly add it to the pipelines build targets.

This is not any kind of best practice for pipeline handlers, but is here
to ensure that this pipeline is always compiled on this branch who's
sole purpose is to use and demonstrate the vivid pipeline handler.