DocumentationUseOnly: Disable compiler warning
Explicitly disable the unused-parameter warning in this pipeline handler.

Parameters are left unused while they are introduced incrementally, so for
documentation purposes only we disable this warning so that we can compile
each commit independently without breaking the flow of the development
additions.

This is not recommended practice within libcamera, please listen to your
compiler warnings.

Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>