fix ControlInfo for Span Controls
Some control properties are typed with a Span to store an array of values.
Currently those are ColourGains, SensorBlackLevels, ColourCorrectionMatrix
and FrameDurationLimits. The value range and defaults in the ControlInfo of
those Controls is currently defined as scalar. This prevents accessing the
ControlInfo via the native Span type.

By defining the ControlInfo in terms of Spans, we can avoid this issue.

Signed-off-by: Christian Rauch <Rauch.Christian@gmx.de>